bucketlist-api/requirements.txt
alembic==0.9.5
bcrypt==3.1.3
blueprint==3.4.2
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
codeclimate-test-reporter==0.2.3
colorama==0.3.9
coverage==4.4.1
first==2.0.1
flake8==3.4.1
flasgger==0.6.6
Flask==0.12.2
Flask-API==1.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Migrate==2.1.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
Flask-Testing==0.6.2
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.7.4
mock==2.0.0
nose==1.3.7
pbr==3.1.1
pew==0.1.26
pip-tools==1.10.0
pipenv==8.1.3
psycopg2==2.7.3.1
py==1.4.34
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.5.0
PyJWT==1.5.3
pytest==3.2.1
python-dateutil==2.6.1
python-editor==1.0.3
pythonz-bd==1.11.4
PyYAML==3.12
requests==2.18.4
resumable-urlretrieve==0.1.5
six==1.10.0
SQLAlchemy==1.1.13
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
Werkzeug==0.12.2
